+++ js/main.jsx
@@ -1,5 +1,4 @@
 require ('../css/main.css');
+require ('./components/list.jsx');
 var React = require ('react');
 var ReactDOM = require ('react-dom');
-
-
